Remove netCDF from requirements.txt
The installation of netCDF on heroku was resulting in an error using
pip.req.parse_requirements (see commit 2b7b71a in this repo). Use
requirements-full.txt or manually install netCDF when using this
package outside of Heroku.  On heroku, just don't use
pyairfire.bluesky.dispersionnc, which depends on netCDF